work around macOS bug
As it turns out, macOS's HID report descriptor parser is not up to spec.
HID 1.11 quote, chapter 6.2.2.8:
Interpretation of Usage, Usage Minimum or Usage Maximum items vary as a
function of the item’s bSize field. If the bSize field = 3 then the item is
interpreted as a 32 bit unsigned value where the high order 16 bits defines the
Usage Page and the low order 16 bits defines the Usage ID. 32 bit usage items
that define both the Usage Page and Usage ID are often referred to as
“Extended” Usages.

The previous report descriptor was recognized correctly by Windows, Linux,
Android at least, but not by macOS.