feat: Allow unit codes without exponent code * Provide `unit::base` which does not have any exponent codes * Allows `unit::unit_item` to be specified with a code as well as a flag, e.g. `unit::unit_type<2>(unit::code::DEGREE)`
This will remove the exponent codes from many of the basic types such as `unit::degree()` and `unit::centimeter()`. It will also clean up the code for simply using `unit::unit_item`, which may be required to exactly match a report descriptor of an existing device by specifying only the unit in a number of bytes without exponent.